requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.28.0
python-dotenv>=1.0.0
